prepare for refactoring

+++ repo_agent/utils/meta_info_utils.py
@@ -67,14 +67,20 @@ def make_fake_files():
                 print(
                     f"{Fore.LIGHTMAGENTA_EX}[Create Temp-File for Deleted(But not Staged) Files]: {Style.RESET_ALL}{now_file_path} -> {latest_file_path}"
                 )
-                with open(
-                    os.path.join(setting.project.target_repo, latest_file_path), "w", encoding='utf-8'
-                ) as writer:
+                try:
+                    with open(
+                        os.path.join(setting.project.target_repo, latest_file_path), "w", encoding='utf-8'
+                    ) as writer:
+                        pass
+                except:
                     pass
-            with open(
-                os.path.join(setting.project.target_repo, now_file_path), "w", encoding='utf-8'
-            ) as writer:
-                writer.write(raw_file_content)
+            try:
+                with open(
+                    os.path.join(setting.project.target_repo, now_file_path), "w", encoding='utf-8'
+                ) as writer:
+                    writer.write(raw_file_content)
+            except:
+                pass
             file_path_reflections[now_file_path] = latest_file_path  # real指向fake
     return file_path_reflections, jump_files
 
@@ -92,7 +98,8 @@ def delete_fake_files():
                 gci(fi_d)
             elif fi_d.endswith(latest_verison_substring):
                 origin_name = fi_d.replace(latest_verison_substring, ".py")
-                os.remove(origin_name)
+                if os.path.exists(origin_name):
+                    os.remove(origin_name)
                 if os.path.getsize(fi_d) == 0:
                     print(
                         f"{Fore.LIGHTRED_EX}[Deleting Temp File]: {Style.RESET_ALL}{fi_d[len(str(setting.project.target_repo)):]}, {origin_name[len(str(setting.project.target_repo)):]}"
